+++ models/memory/arraymemory.h
@@ -44,15 +44,16 @@
 #ifndef ARRAYMEMORY_H
 #define ARRAYMEMORY_H
 
-#include "verbose.h"
 #include "memdevice.h"
-#include "ext_erase.h"
 #include "vmap.h"
 
-#include <greensocket/target/single_socket.h>
 #include <greencontrol/config.h>
-#include <systemc.h>
+#include <greensocket/target/single_socket.h>
+#include "ext_erase.h"
 #include <tlm.h>
+#include <systemc.h>
+
+#include "verbose.h"
 
 /// @brief This class models a array memory. Depending on the configuration
 /// it can be used as ROM, IO, SRAM or SDRAM, in conjunction with the SoCRocket MCTRL.
@@ -113,10 +114,10 @@ class ArrayMemory : public MEMDevice,
         gs::gs_param_array m_performance_counters;
 
         /// Performance counter to store transaction byte reads
-        gs::gs_param<uint64_t> m_reads;
+        gs::gs_param<unsigned long long> m_reads;
 
         /// Performance counter to store the transaction byte writes
-        gs::gs_param<uint64_t> m_writes; 
+        gs::gs_param<unsigned long long> m_writes; 
 };
 
 #endif
