+++ models/mmu_cache/lib/mmu.h
@@ -48,9 +48,9 @@
 #ifndef __MMU_H__
 #define __MMU_H__
 
-#include <tlm.h>
 #include <map>
 #include <greencontrol/config.h>
+#include <tlm.h>
 
 #include "mmu_if.h"
 #include "tlb_adaptor.h"
@@ -302,12 +302,12 @@ class mmu : public sc_core::sc_module, public mmu_if {
   gs::gs_param_array m_performance_counters;
 
 	/// Number of TLB hits
-  gs::gs_param<uint64_t *> tihits;
-  gs::gs_param<uint64_t *> tdhits;
+  gs::gs_param<unsigned long long *> tihits;
+  gs::gs_param<unsigned long long *> tdhits;
 
 	/// Number of TLB misses
-  gs::gs_param<uint64_t> timisses;
-  gs::gs_param<uint64_t> tdmisses;
+  gs::gs_param<unsigned long long> timisses;
+  gs::gs_param<unsigned long long> tdmisses;
 
 	/// Pseudo random counter for LRU
 	uint32_t m_pseudo_rand;
